requirements.txt
fastapi
sqlalchemy
mysql-connector-python
python-dotenv
pydantic